refact: SocketSendData(), ListenSocket and NODE_ADDR structs moved to net.h
